requirements.txt
# Telegram Bot
python-telegram-bot==21.0.1

# Web Framework
fastapi==0.110.0
uvicorn==0.28.0
jinja2==3.1.4
python-multipart==0.0.5

# Video Processing
yt-dlp==2025.10.14
Pillow>=10.0.0

# Text-to-Speech (Edge TTS)
edge-tts>=6.1.0

# Database
supabase>=2.0.0

# HTTP Client
httpx==0.27.0
aiofiles==23.2.1
requests>=2.31.0

# Environment
python-dotenv==1.0.1
certifi==2024.2.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.0
pytest-cov==4.0.0
